chore(trello): rename webhook files to kebab-case
